core+gtk: align session-duration seconds on u32 across modes
`TIMER_DEFAULT_SECS` was `u64`; `BREATHING_DEFAULT_SECS` was
`u32`. Same kind of value (a session-length in seconds), different
widths, forcing the shell to scatter `as u64` / `as u32` casts
through every place that touched both. The backlog convention is
"u32 for a single session duration, u64 only where
`Duration::as_secs()` feeds the value, i64 for DB-aggregated
totals from chrono."

Aligned the Timer path with the breathing path on u32:

  - `meditate_core::session::TIMER_DEFAULT_SECS` u64 → u32
  - `format::format_hhmm(secs: u64)` → `secs: u32`
  - `format::idle_hero_label(target_secs: u64)` → `target_secs: u32`
  - shell's `countdown_target_secs: Cell<u64>` → `Cell<u32>`
  - `set_countdown_target(secs: u64)` → `set_countdown_target(secs: u32)`

Cast accounting at call sites in `src/timer/imp.rs`:

  - Dropped 6 explicit `as u32` / `as u64` casts that the
    type-width mismatch had forced (countdown ↔ breathing,
    countdown ↔ format helpers).
  - Added 1 `as u64` cast where countdown still has to feed
    `Duration::from_secs(...)` (`tick_overtime` hero refresh) —
    Duration's API is u64, no way around that.
  - Added 1 `as u64` cast on the overtime fallback path where
    `core_session_end` returns `Option<u64>` and the fallback is
    the planned-duration target.

Net: -4 casts at the call sites, plus the constants now actually
match what the backlog said the convention was. No behaviour
change.

+++ meditate-core/src/format.rs
@@ -188,7 +188,7 @@ pub fn ellipsize(s: &str, max_chars: usize) -> String {
 /// duration in `HH:MM`. The shell resolves the right `target_secs`
 /// per mode (Timer's countdown / Box-Breath's session length /
 /// Guided's probed file length) and passes it in.
-pub fn idle_hero_label(stopwatch_on: bool, target_secs: u64) -> String {
+pub fn idle_hero_label(stopwatch_on: bool, target_secs: u32) -> String {
     if stopwatch_on {
         "00:00".to_string()
     } else {
@@ -211,7 +211,7 @@ pub fn box_breath_counter_label(elapsed: Duration, target: Option<Duration>) -> 
 /// Used by the Setup view's hero label and the Duration row's value
 /// suffix where seconds aren't shown (minute-aligned by the spinner).
 /// Hours zero-pad to two digits to keep the label width stable.
-pub fn format_hhmm(secs: u64) -> String {
+pub fn format_hhmm(secs: u32) -> String {
     let h = secs / 3600;
     let m = (secs % 3600) / 60;
     format!("{h:02}:{m:02}")
